First new commit
Had to nuke whole repo and create a new one because I inadvertently uploaded a version with one single word of sensitive info in it, and GitHub, if you make a single tiny mistake, intelligently gives you absolutely no way to delete a commit or remove it from eternal public view without nuking the repo, its entire version history, and all prior associations with the project. It's seriously great design.